+++ src/common/mod.rs
@@ -1,2 +1,3 @@
+pub mod fmt;
 pub mod numeric;
 pub mod sealed;