+++ src/lib.rs
@@ -65,8 +65,6 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 #![cfg_attr(docsrs, feature(doc_cfg))]
 
-extern crate alloc;
-
 pub mod ago;
 pub mod bytes;
 #[cfg(feature = "chrono")]